pl compiler: refactor into smaller files
